backend/requirements.txt
# Core dependencies
setuptools>=68.0.0
fastapi>=0.104.1
uvicorn[standard]>=0.24.0
sqlalchemy>=2.0.23
pydantic>=2.5.0
python-dotenv>=1.0.0

# Database
psycopg2-binary>=2.9.9

# ZIP processing
watchdog>=3.0.0

# Utilities
python-magic>=0.4.27

# Telegram integration
telethon>=1.41.0

# Note: MEGA downloads use system 'megatools' package
# Install with: sudo apt install megatools
